Backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
python-dotenv==1.0.0
flask-talisman==1.1.0
flask-limiter==3.5.0
flask-caching==2.1.0
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23
werkzeug==3.0.1
gevent==23.9.1
gevent-websocket==0.10.1
